Added navigation to post detail page from profile and posts fragments + misc fixes

+++ app/src/main/java/com/jazart/symphony/posts/PostsViewModel.kt
@@ -46,8 +46,8 @@ class PostsViewModel : BaseViewModel() {
         }
     }
 
-    fun deletePost(postId: String) {
-        firebaseRepo.deletePost(postId)
+    fun deletePost(postId: String?) {
+        if(postId != null) firebaseRepo.deletePost(postId)
         refreshContent()
     }
